hotfix: use api instead of mqtt for sending messages (#173)
As reported by several users, we have now identified and are
investigating a phenomenon named "Schodinger's messages", namely when
messages sent by the client via MQTT mainly is unstable and might not be
received by other web clients. We suspect we are using an outdated MQTT
endpoint so that the database is not syncing and this PR currently
reverts to using API instead of realtime client for sending messages.
See #171

**This PR ONLY fixes issues with sending text, relies and reactions will
be further investigated**