[Feature] Viewing Stories (#179)
## Summary

This PR closes #87 by adding a `instagram-cli stories` command that
renders the new stories view. The behaviour should closely mimic
Instagram web / mobile. **This feature is experimental, see the first
comment below** _Bumps version to 1.3.0._

<img width="1171" height="886" alt="Screenshot 2025-11-22 at 10 43
55 pm"
src="https://github.com/user-attachments/assets/8e8022bf-4a91-47b9-81c8-feae4927429c"
/>

## Data flow

This happens after running the command:

- Use the `reelsTray` endpoint for recommended stories from Meta, this
builds the initial unread stories list (**this list is usually extremely
large!** For example, on my personal account this is nearly 100 items!
It contains all the unread stories that your account has, fetching
everything at once is suicidal 💀 ). `reelsTray` is quite lightweight
because it only contains "metadata" instead of the actual story content
(as opposed to the documentation 😭), we use this to get a list of users
whom we need to check, this is ALSO ranked (it returns a list of ranked
scores)
- The `useStories` hook will then fetch and populate one story, then
prefetch two more using (not awaited) the `getStoryFromUser` function
that calls the `userStory` endpoint, this returns the actual story
content and we cache it by userID so navigating back is fast
- Lazy loading is implemented by returning a `loadMore(index)` callback
from the hook, this loads the next 3 stories but two of them should be
prefetched already so it load the 3rd story from current index.
Prefetching will help reduce loading time
- The Story View will call `loadMore` whenever the user navigates the
list of stories, we group story items by user to display it like a
Carousel. Meanwhile, we send read status for the story by building a
local set of read story IDs, this operation is not awaited.
- `s` key enters Search Mode, where the user types an username (**NOT
display name**) and we search for the corresponding userId then stories
for that user, which gets prepended to the top of the list
- Videos can be opened in the browser as usual

Unironically, the `reelsTray` endpoint name very well describes what
Instagram does and we shall illustrate it with a little bit of AI slop
here. Another piece of daily motivation to migrate to Instagram CLI and
save yourself 😄:

<img width="512" height="512" alt="unknown"
src="https://github.com/user-attachments/assets/65f426ee-339c-4b09-92e3-d9e60083ca3b"
/>

## Type Definitions

- Story is an object that carries image and/or video and metadata
- `StoryReel` is a "Story" that you see as a circle object that you can
click open, it associates a User with a list of Stories
- `ReelMentions` is `Users[]` that indicate the users being mentioned in
the story, monkey patched!

## Chore

- Moved all logger usage to outside of the component and removed all
logger references from useEffect dep list
- Renamed and combined a few CI, made sure that Typescript CI runs on
pushes to main
- Removed all `console.xxx` statements and turned on warning in the
linter, use `logger` whenever possible!
- Refactored media components -- media pane and Split View, shared them
over feed and the new stories view