[Feature] search for threads in chat menu and CLI (#198)
Add search functionality in chat menu and CLI to the TypeScript client
(which is already present in the python client).

## Features
- In CLI:
  ```
  $instagram-cli --help
  Usage: cli chat [options] [username]

  Arguments:
    username                   Username to login with (optional)

  Options:
-t, --title [title] Search for a chat by title. If a match is found,
directly enter the chat.
-u, --username [username] Search for a chat by username. If a match is
found, directly enter the chat.
    -h, --help                 Show help
  ```
- When searching with the CLI and a exact match is found for username,
or a high similarity match is found for title, directly enters the
messaging view of the thread in question. Otherwise, the CLI query is
filled-in in the search input in the TUI.
- In chat menu, press `/` or `@` to search by title or username
respectively. Use Esc to quit search input. Search results will be
displayed as a threads list. Use arrow keys to move up / down in the
search results. Searching has a 300ms debounce.

## Backend
- Added `searchThreadsByUsername()` to `client.ts`, which directly
searches for the exact username query.
- Added `searchThreadsByTitle()` to `client.ts`, which uses `Fuse.js` to
perform a fuzzy search on the titles of the user's threads, returning a
list of threads along with their similarity score with the query string.
- Threads are now cached in the client's `threadsCache`, so each search
doesn't need to re-fetch all threads and bombard Meta's API

## To-do / technical debt
- Implement search-related changes in the mock client (DONE)

EDIT: Added error message to guide users for bad password issue see #199

---------

Co-authored-by: Jet Chiang <jetjiang.ez@gmail.com>